bug rating >>> review amount

+++ app/src/main/java/com/review/foodreview/SearchBarFragment.java
@@ -115,7 +115,7 @@ public class SearchBarFragment extends Fragment{
                     (HashMap<String, Long>) result.get("rating"),
                     (List<String>) result.get("imageUri"),
                     (List<DocumentReference>) result.get("review"),
-                    0
+                    (long) 0
             );
             listForAdapter.add(restaurant);
             if (getContext() != null) {
